requirements.txt

# Requirements for ai-langchain
# Pinned / minimum versions that are known to work in the dev environment.
langchain==1.1.2
openai>=2.9.0
python-dotenv>=1.0.0
langchain-anthropic>=0.3.0
chromadb>=1.3.5
tiktoken>=0.12.0
fastapi>=0.95.0
uvicorn[standard]>=0.18.3
rich>=14.0.0
